+++ src/codes/rudolph/idea/cfg/CFGIcons.java
@@ -4,6 +4,6 @@ import com.intellij.openapi.util.IconLoader;
 
 import javax.swing.Icon;
 
-public class Icons {
+public class CFGIcons {
     public static final Icon FILE = IconLoader.getIcon("/codes/rudolph/idea/cfg/jar-gray.png");
 }